Support new format of config file input, while creating connector from ccloud cli
Currently the format of the config file across several channels(CLI, API) is not consistent.
When user tries to create a connector from cloud UI, it also gives an output in JSON format before creating a connector.
So, we have 3 different places, each following a different convention.

As part of this PR, enhancing the ccloud cli to accept the API format, in addition to the currently accepted format.

A sample input:
---------------
{
 "name" : "DatagenConnector",
  "config" : {
     "name": "DatagenConnector",
     "key1" : "val1",
     "key2" : "val2"
  }
}